Day 5: make stack top correspond to ArrayList end
Previously, the top of each stack was the first item in each ArrayList.
This meant removing the first element and inserting it before the first
element each time a crate was moved. These operations are both O(N),
whereas popping and pushing (appending in Zig ArrayList terminology) are
O(1).

Even though the part 2 solution requires orderedRemove() no matter what,
we still get to replace an insert() with an append() which saves time.